backend/requirements.txt
Django==5.2.8
python-dotenv==1.0.0
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.3
requests==2.32.3
